+++ src/components/HarvestingConfiguration/AggregatorInfo/index.js
@@ -1,7 +1,0 @@
-import AggregatorInfoForm from './AggregatorInfoForm';
-import AggregatorInfoView from './AggregatorInfoView';
-
-export {
-  AggregatorInfoForm,
-  AggregatorInfoView,
-};