+++ src/components/UDPFilters/UDPFilters.js
@@ -35,7 +35,6 @@ const UDPFilters = ({
   const [filterState, setFilterState] = useState({
     harvestingStatus: [],
     harvestVia: [],
-    aggregators: [],
     hasFailedReport: [],
     tags: [],
     errorCodes: [],
@@ -183,7 +182,6 @@ const UDPFilters = ({
       {renderCheckboxFilter('status')}
       {renderCheckboxFilter('harvestingStatus')}
       {renderCheckboxFilter('harvestVia')}
-      {renderCheckboxFilter('aggregators', true)}
       {renderMultiSelectionFilter('reportTypes')}
       {renderMultiSelectionFilter('reportReleases')}
       {renderCheckboxFilter('hasFailedReport', true)}