+++ src/components/views/UDPForm.test.js
@@ -14,29 +14,6 @@ import {
 import renderWithIntl from '../../../test/jest/helpers/renderWithIntl';
 import UDPForm from './UDPForm';
 
-const stubAggregators = [
-  {
-    aggregatorSettings: [
-      {
-        id: '5b6ba83e-d7e5-414e-ba7b-134749c0d950',
-        label: 'German National Statistics Server',
-        serviceUrl: 'https://sushi.url-to-nss.de/Sushiservice/GetReport',
-        serviceType: 'NSS',
-        accountConfig: {
-          configMail: 'accounts@example.org',
-          configType: 'Mail',
-          displayContact: ['John Doe, Phone +49 0000 0000000 '],
-        },
-        aggregatorConfig: {
-          apiKey: 'xxx',
-          customerId: 'xxx',
-          requestorId: 'xxx',
-        },
-      },
-    ],
-  },
-];
-
 const stubHarvesterImpls = [
   {
     implementations: [
@@ -115,7 +92,6 @@ const renderUDPForm = (stripes, udp = initialValues, harvesterImpls = stubHarves
       <MemoryRouter>
         <UDPForm
           data={{
-            aggregators: stubAggregators,
             harvesterImpls,
           }}
           handlers={{ onClose }}
@@ -189,17 +165,15 @@ describe('UDPForm', () => {
       await userEvent.selectOptions(harvestingStatusSelect, 'active');
     });
 
-    test('should enable aggregator options', async () => {
-      await userEvent.selectOptions(screen.getByRole('combobox', { name: /harvest statistics via/i }), ['aggregator']);
+    test('should not offer aggregator option', () => {
+      const harvestViaSelect = screen.getByRole('combobox', { name: /harvest statistics via/i });
 
-      expect(screen.getByRole('combobox', { name: 'Aggregator' })).toBeEnabled();
-      expect(screen.getByRole('combobox', { name: 'Service type' })).toBeDisabled();
+      expect(within(harvestViaSelect).queryByRole('option', { name: 'Aggregator' })).not.toBeInTheDocument();
     });
 
     test('should enable sushi options', async () => {
       await userEvent.selectOptions(screen.getByRole('combobox', { name: /harvest statistics via/i }), ['sushi']);
 
-      expect(screen.getByRole('combobox', { name: 'Aggregator' })).toBeDisabled();
       expect(screen.getByRole('combobox', { name: 'Service type' })).toBeEnabled();
     });
   });
@@ -331,11 +305,10 @@ describe('UDPForm', () => {
       await userEvent.type(screen.getByRole('textbox', { name: /provider name/i }), 'FooBar');
       await userEvent.selectOptions(screen.getByRole('combobox', { name: /provider status/i }), 'active');
       await userEvent.selectOptions(screen.getByRole('combobox', { name: /harvesting status/i }), 'active');
+      await userEvent.selectOptions(screen.getByRole('combobox', { name: /harvest statistics via/i }), ['sushi']);
       await userEvent.selectOptions(screen.getByRole('combobox', { name: /service type/i }), ['Counter-Sushi 4.1']);
-      await userEvent.selectOptions(screen.getByRole('combobox', { name: /harvest statistics via/i }), ['aggregator']);
-      await userEvent.selectOptions(
-        screen.getByRole('combobox', { name: /aggregator/i }), ['5b6ba83e-d7e5-414e-ba7b-134749c0d950']
-      );
+      await userEvent.type(screen.getByRole('textbox', { name: /service url/i }), 'http://abc');
+      await userEvent.type(screen.getByRole('textbox', { name: /customer id/i }), 'MyCustomerID');
       await userEvent.click(screen.getByRole('button', { name: /add report type/i }));
 
       const reportTypeButton = screen.getByRole('button', { name: 'Report type' });
@@ -351,32 +324,6 @@ describe('UDPForm', () => {
     });
   });
 
-  describe('test change from sushi to aggregator', () => {
-    beforeEach(() => {
-      renderUDPForm(stripes);
-    });
-
-    test('form is invalid when changing from sushi to aggregator', async () => {
-      await userEvent.type(screen.getByRole('textbox', { name: /provider name/i }), 'FooBar');
-      await userEvent.selectOptions(screen.getByRole('combobox', { name: /harvesting status/i }), 'active');
-      await userEvent.selectOptions(screen.getByRole('combobox', { name: /harvest statistics via/i }), ['sushi']);
-      await userEvent.selectOptions(screen.getByRole('combobox', { name: /service type/i }), ['Counter-Sushi 4.1']);
-      await userEvent.type(screen.getByRole('textbox', { name: /service url/i }), 'http://abc');
-
-      await userEvent.click(screen.getByRole('button', { name: /add report type/i }));
-
-      const reportTypeButton = screen.getByRole('button', { name: 'Report type' });
-      expect(reportTypeButton).toBeInTheDocument();
-      await userEvent.click(reportTypeButton);
-      await userEvent.click(screen.getByText('BR1'));
-      await userEvent.type(screen.getByLabelText(/Harvesting start/i), '01/2020');
-      await userEvent.type(screen.getByRole('textbox', { name: /customer id/i }), 'MyCustomerID');
-      await userEvent.selectOptions(screen.getByRole('combobox', { name: /harvest statistics via/i }), ['aggregator']);
-      await userEvent.click(screen.getByRole('button', { name: 'Save & close' }));
-      expect(onSubmit).not.toHaveBeenCalled();
-    });
-  });
-
   describe('test required attributes for harvesting status choice', () => {
     beforeEach(() => {
       renderUDPForm(stripes);
@@ -389,8 +336,6 @@ describe('UDPForm', () => {
       expect(screen.getByRole('textbox', { name: 'Description' })).not.toBeRequired();
       expect(screen.getByRole('combobox', { name: 'Harvesting status' })).toBeRequired();
       expect(screen.getByRole('combobox', { name: 'Harvest statistics via' })).toBeRequired();
-      expect(screen.getByRole('combobox', { name: 'Aggregator' })).not.toBeRequired();
-      expect(screen.getByRole('textbox', { name: 'Vendor code' })).not.toBeRequired();
 
       expect(screen.getByRole('combobox', { name: 'Service type' })).toBeRequired();
       expect(screen.getByRole('textbox', { name: /service url/i })).toBeRequired();
@@ -418,8 +363,6 @@ describe('UDPForm', () => {
       expect(screen.getByRole('textbox', { name: 'Description' })).not.toBeRequired();
       expect(screen.getByRole('combobox', { name: 'Harvesting status' })).toBeRequired();
       expect(screen.getByRole('combobox', { name: 'Harvest statistics via' })).not.toBeRequired();
-      expect(screen.getByRole('combobox', { name: 'Aggregator' })).not.toBeRequired();
-      expect(screen.getByRole('textbox', { name: 'Vendor code' })).not.toBeRequired();
 
       expect(screen.getByRole('combobox', { name: 'Service type' })).not.toBeRequired();
       expect(screen.getByRole('textbox', { name: /service url/i })).not.toBeRequired();
@@ -459,25 +402,6 @@ describe('UDPForm', () => {
       expect(providerNameTextbox).toHaveValue('FooBar');
       await userEvent.click(saveAndCloseButton);
       expect(onSubmit).toHaveBeenCalled();
-    });
-
-    describe('test required value of customerId field', () => {
-      test('change harvest statistics via from sushi to aggregator', async () => {
-        await userEvent.selectOptions(screen.getByRole('combobox', { name: /harvesting status/i }), 'active');
-        await userEvent.selectOptions(screen.getByRole('combobox', { name: /harvest statistics via/i }), ['sushi']);
-        await userEvent.click(screen.getByRole('textbox', { name: 'Customer ID' }));
-        await userEvent.click(screen.getByRole('textbox', { name: 'Platform' }));
-
-        expect(screen.getByText('Required')).toBeInTheDocument();
-        expect(screen.getByRole('textbox', { name: 'Customer ID' })).toBeRequired();
-
-        await userEvent.selectOptions(
-          screen.getByRole('combobox', { name: /harvest statistics via/i }),
-          ['aggregator']
-        );
-        expect(screen.getByRole('textbox', { name: 'Customer ID' })).not.toBeRequired();
-        expect(screen.queryByText('Required')).not.toBeInTheDocument();
-      });
     });
   });